# api.spec.yaml
openapi: 3.0.0

info:
  description: |
    This is the expense tracking app
  version: 1.0.0
  title: Expense

tags:
  - name: Groups
    description: Group management APIs
  - name: Auth
    description: Authentication APIs

paths:
  /api/auth/login:
    post:
      security: []
      tags:
        - Auth
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ILogin"
      responses:
        "200":
          description: Success
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
                example: session_id=abc123; HttpOnly; Secure; SameSite=Lax; Path=/secur
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/groups:
    get:
      tags:
        - Groups
      operationId: getMyGroups
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupsList"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Groups
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IGroup"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/groups/{id}:
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: integer
    get:
      tags:
        - Groups
      operationId: getGroup
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags:
        - Groups
      operationId: updateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IGroup"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Groups
      operationId: deleteGroup
      responses:
        "204":
          description: Success
        "409":
          description: Conflict. Some Condition failed
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id

  schemas:
    ILogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    GroupsList:
      type: array
      items:
        $ref: "#/components/schemas/Group"

    Group:
      allOf:
        - $ref: "#/components/schemas/IGroup"
        - $ref: "#/components/schemas/Id"

    IGroup:
      type: object
      required:
        - name
      properties:
        name:
          type: string

    Id:
      type: object
      required:
        - id
      properties:
        id:
          type: integer

  responses:
    BadRequest:
      description: Bad request or input
    InternalServerError:
      description: Something went wrong with the server

security:
  - sessionCookie: []